refactor: TempDirGuard をクレートごとの testutil モジュールに共通化
5箇所に重複していた TempDirGuard を 2 つの共通モジュールに集約:
- crates/ghost-meta/src/testutil.rs（descript, ghost, thumbnail の 3 箇所分）
- src-tauri/src/testutil.rs（commands/ghost/mod, commands/locale の 2 箇所分）

API を panic 方式（-> Self）に統一し、呼び出し元の不要な .unwrap() / ? を除去。
