+++ crates/ghost-meta/src/testutil.rs
@@ -1,0 +1,26 @@
+/// テスト用の一時ディレクトリ。Drop 時に自動削除される。
+pub(crate) struct TempDirGuard {
+    path: std::path::PathBuf,
+}
+
+impl TempDirGuard {
+    pub(crate) fn new(prefix: &str) -> Self {
+        let now = std::time::SystemTime::now()
+            .duration_since(std::time::UNIX_EPOCH)
+            .unwrap()
+            .as_nanos();
+        let path = std::env::temp_dir().join(format!("{}_{}", prefix, now));
+        std::fs::create_dir_all(&path).unwrap();
+        Self { path }
+    }
+
+    pub(crate) fn path(&self) -> &std::path::PathBuf {
+        &self.path
+    }
+}
+
+impl Drop for TempDirGuard {
+    fn drop(&mut self) {
+        let _ = std::fs::remove_dir_all(&self.path);
+    }
+}
