+++ crates/ghost-meta/src/thumbnail.rs
@@ -154,34 +154,8 @@ fn detect_thumbnail_alpha(_path: &Path) -> AlphaMode {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::testutil::TempDirGuard;
     use std::path::PathBuf;
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    struct TempDirGuard {
-        path: PathBuf,
-    }
-
-    impl TempDirGuard {
-        fn new(prefix: &str) -> Self {
-            let now = SystemTime::now()
-                .duration_since(UNIX_EPOCH)
-                .unwrap()
-                .as_nanos();
-            let path = std::env::temp_dir().join(format!("{}_{}", prefix, now));
-            fs::create_dir_all(&path).unwrap();
-            Self { path }
-        }
-
-        fn path(&self) -> &PathBuf {
-            &self.path
-        }
-    }
-
-    impl Drop for TempDirGuard {
-        fn drop(&mut self) {
-            let _ = fs::remove_dir_all(&self.path);
-        }
-    }
 
     fn create_shell_master(ghost_root: &PathBuf) -> PathBuf {
         let shell_master = ghost_root.join("shell").join("master");
