+++ src-tauri/src/commands/ghost/mod.rs
@@ -28,37 +28,9 @@ pub fn get_ghosts_fingerprint(
 mod tests {
     use super::fingerprint::build_fingerprint;
     use super::scan::{scan_ghosts_with_fingerprint_internal, unique_sorted_additional_folders};
+    use crate::testutil::TempDirGuard;
     use std::fs;
     use std::path::PathBuf;
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    struct TempDirGuard {
-        path: PathBuf,
-    }
-
-    impl TempDirGuard {
-        fn new(prefix: &str) -> Result<Self, String> {
-            let now = SystemTime::now()
-                .duration_since(UNIX_EPOCH)
-                .map_err(|error| format!("failed to get time: {}", error))?
-                .as_nanos();
-            let path = std::env::temp_dir().join(format!("{}_{}", prefix, now));
-            fs::create_dir_all(&path).map_err(|error| {
-                format!("failed to create temp dir {}: {}", path.display(), error)
-            })?;
-            Ok(Self { path })
-        }
-
-        fn path(&self) -> &PathBuf {
-            &self.path
-        }
-    }
-
-    impl Drop for TempDirGuard {
-        fn drop(&mut self) {
-            let _ = fs::remove_dir_all(&self.path);
-        }
-    }
 
     fn create_ghost_dir(root: &PathBuf, name: &str) -> Result<(), String> {
         create_ghost_dir_with_descript(root, name, "name,Test Ghost\ncharset,UTF-8\n")
@@ -93,7 +65,7 @@ mod tests {
 
     #[test]
     fn build_fingerprint_is_order_independent_for_additional_folders() -> Result<(), String> {
-        let workspace = TempDirGuard::new("ghost_launcher_fingerprint_test")?;
+        let workspace = TempDirGuard::new("ghost_launcher_fingerprint_test");
         let ssp_root = workspace.path().join("ssp");
         let ssp_ghost = ssp_root.join("ghost");
         fs::create_dir_all(&ssp_ghost)
@@ -127,7 +99,7 @@ mod tests {
 
     #[test]
     fn scan_ghosts_internal_collects_sources_and_sorts_by_name() -> Result<(), String> {
-        let workspace = TempDirGuard::new("ghost_launcher_scan_test")?;
+        let workspace = TempDirGuard::new("ghost_launcher_scan_test");
         let ssp_root = workspace.path().join("ssp");
         let ssp_ghost = ssp_root.join("ghost");
         fs::create_dir_all(&ssp_ghost)
@@ -178,7 +150,7 @@ mod tests {
 
     #[test]
     fn scan_ghosts_internal_extracts_craftman_field() -> Result<(), String> {
-        let workspace = TempDirGuard::new("ghost_launcher_craftman_test")?;
+        let workspace = TempDirGuard::new("ghost_launcher_craftman_test");
         let ssp_root = workspace.path().join("ssp");
         let ssp_ghost = ssp_root.join("ghost");
         fs::create_dir_all(&ssp_ghost)
@@ -214,7 +186,7 @@ mod tests {
     #[test]
     fn scan_ghosts_internal_falls_back_to_directory_name_without_name_field() -> Result<(), String>
     {
-        let workspace = TempDirGuard::new("ghost_launcher_scan_fallback_test")?;
+        let workspace = TempDirGuard::new("ghost_launcher_scan_fallback_test");
         let ssp_root = workspace.path().join("ssp");
         let ssp_ghost = ssp_root.join("ghost");
         fs::create_dir_all(&ssp_ghost)
@@ -237,7 +209,7 @@ mod tests {
 
     #[test]
     fn scan_ghosts_internal_returns_error_when_ssp_ghost_dir_is_missing() -> Result<(), String> {
-        let workspace = TempDirGuard::new("ghost_launcher_missing_ghost_dir_test")?;
+        let workspace = TempDirGuard::new("ghost_launcher_missing_ghost_dir_test");
         let ssp_root = workspace.path().join("ssp_without_ghost");
         fs::create_dir_all(&ssp_root)
             .map_err(|error| format!("failed to create ssp root dir: {}", error))?;
@@ -251,7 +223,7 @@ mod tests {
 
     #[test]
     fn fingerprint_with_missing_additional_folder_matches_scan_fingerprint() -> Result<(), String> {
-        let workspace = TempDirGuard::new("ghost_launcher_missing_folder_fp_test")?;
+        let workspace = TempDirGuard::new("ghost_launcher_missing_folder_fp_test");
         let ssp_root = workspace.path().join("ssp");
         let ssp_ghost = ssp_root.join("ghost");
         fs::create_dir_all(&ssp_ghost)
@@ -280,7 +252,7 @@ mod tests {
 
     #[test]
     fn integrated_fingerprint_matches_standalone_build_fingerprint() -> Result<(), String> {
-        let workspace = TempDirGuard::new("ghost_launcher_fp_consistency_test")?;
+        let workspace = TempDirGuard::new("ghost_launcher_fp_consistency_test");
         let ssp_root = workspace.path().join("ssp");
         let ssp_ghost = ssp_root.join("ghost");
         fs::create_dir_all(&ssp_ghost)
